parent delete

+++ api_components/employee.js
@@ -1,3 +1,4 @@
+
 // flow
 var express = require("express");
 var config = require("../config.json");
@@ -585,11 +586,21 @@ router.route('/edit_employee_details/:employee_id')
         var blood_group = req.body.blood_group;
         var spoken_languages = req.body.spoken_languages;
         var country = req.body.country;
+        var state = req.body.state;
+        var alternate_email = req.body.alternate_email;
         var perm_city = req.body.perm_city;
         var postal_code = req.body.postal_code;
+        var perm_address = req.body.perm_address;
+        var cur_address = req.body.cur_address;
+        var martial_status = req.body.martial_status
+
+
+
+
 
         mongo.connect(url, function (err, db) {
             db.collection('employee').update(myquery, {
+
                 $set: {
                     first_name: req_first_name,
                     last_name: req_last_name,
@@ -599,16 +610,24 @@ router.route('/edit_employee_details/:employee_id')
                     job_category: req_job_category,
                     experience: req_experience,
                     phone: req_phone,
-                    email: req_email,                    
+                    email: req_email,
                     website: req_website,
                     joined_on: req_joined_on,
                     salary_band: salary_band,
+                    alternate_email: alternate_email,
+                    state: state,
                     basic_pay: basic_pay,
+                    //  cur_address: cur_address,
+                    // perm_address: perm_address,
                     mobile: mobile,
                     blood_group: blood_group,
                     spoken_languages: spoken_languages,
                     country: country,
-                    postal_code: postal_code
+                    postal_code: postal_code,
+                    //   perm_city: perm_city,
+                    martial_status: martial_status,
+                    permanent_address: [{ perm_city: perm_city, perm_address: perm_address }],
+                    current_address: [{ cur_address: cur_address }]
                 }
             }, function (err, result) {
                 assert.equal(null, err);
@@ -621,7 +640,34 @@ router.route('/edit_employee_details/:employee_id')
         });
     });
 
+// else {
+//     mongo.connect(url, function (err, db) {
+//         var cursor = db.collection('teachers').find({ employee_id: employee_id });
+//         cursor.forEach(function (doc, err) {
+//             assert.equal(null, err);
+//             resultArray.push(doc);
+//         }, function () {
+//             uniqueId = resultArray[0].teacher_id;
+//             db.collection('teachers').deleteOne(myquery, function (err, result) {
+//                 if (err) {
+//                     res.send('teacher false');
+//                 }
+//                 else {
+//                     mongo.connect(loginUrl, function (err, db) {
+//                         db.collection('users').deleteOne({ uniqueId: uniqueId }, function (err, result) {
+//                             assert.equal(null, err);
+//                             if (err) {
+//                                 res.send('user false');
+//                             }
+//                         });
+//                     });
+//                 }
 
+//             });
+
+//         });
+//     });
+// }
 
 router.route('/delete_employee/:employee_id')
     .delete(function (req, res, next) {
@@ -642,34 +688,54 @@ router.route('/delete_employee/:employee_id')
                             assert.equal(null, err);
                             resultArray.push(doc);
                         }, function () {
-                            uniqueId = resultArray[0].teacher_id;
-                            db.collection('teachers').deleteOne(myquery, function (err, result) {
-                                if (err) {
-                                    res.send('teacher false');
-                                }
-                                else {
-                                    mongo.connect(loginUrl, function (err, db) {
-                                        db.collection('users').deleteOne({ uniqueId: uniqueId }, function (err, result) {
-                                            assert.equal(null, err);
-                                            if (err) {
-                                                res.send('user false');
-                                            }
+                          //  console.log(resultArray);
+                            length = resultArray.length;
+                         //   console.log(length);
+                            if (length != 0) {
+                                uniqueId = resultArray[0].teacher_id;
+                                db.collection('teachers').deleteOne(myquery, function (err, result) {
+                                    if (err) {
+                                        res.send('teacher false');
+                                    }
+                                    else {
+                                        mongo.connect(loginUrl, function (err, db) {
+                                            db.collection('users').deleteOne({ uniqueId: uniqueId }, function (err, result) {
+                                                assert.equal(null, err);
+                                                if (err) {
+                                                    res.send('user false');
+                                                }
+                                            });
                                         });
-                                    });
-                                }
+                                    }
 
-                            });
+                                });
+                            }
 
                         });
                     });
                 }
-                db.close();
+                db.close(); 
                 res.send('true');
             });
 
         });
     });
 
+// router.route('/delete_employee/:employee_id')
+// .delete(function (req, res, next) {
+//     var myquery = { employee_id: req.params.employee_id };
+
+//     mongo.connect(url, function (err, db) {
+//         db.collection('employee').deleteOne(myquery, function (err, result) {
+//             assert.equal(null, err);
+//             if (err) {
+//                 res.send('false');
+//             }
+//             db.close();
+//             res.send('true');
+//         });
+//     });
+// });
 
 router.route('/employee_photo_edit/:employee_id')
     .post(function (req, res, next) {
